mandelbrot.c: refactor to split cli/main from mandelbrot calculation
